net_slist: update types, declarations
Mostly this changes array indices and externally-sourced data that
should never trigger signed overflow to unsigned,
and changes to lhnetaddresstype_t where applicable.

Fixes an old bug in the DP Quake server list where if there are less
servers than the most that could fit onscreen, the list jumps around
and/or duplicates entries when moving the cursor down.

+++ mvm_cmds.c
@@ -467,7 +467,7 @@ static void VM_M_getserverliststring(prvm_prog_t *prog)
 	}
 	else
 	{
-		if(hostnr < 0 || hostnr >= serverlist_viewcount)
+		if(hostnr < 0 || (unsigned)hostnr >= serverlist_viewcount)
 		{
 			Con_Print("VM_M_getserverliststring: bad hostnr passed!\n");
 			return;
@@ -532,7 +532,7 @@ static void VM_M_getserverlistnumber(prvm_prog_t *prog)
 	}
 	else
 	{
-		if(hostnr < 0 || hostnr >= serverlist_viewcount)
+		if(hostnr < 0 || (unsigned)hostnr >= serverlist_viewcount)
 		{
 			Con_Print("VM_M_getserverliststring: bad hostnr passed!\n");
 			return;
